requirements.txt
Django==4.0.4
django-cors-headers==3.8.0
djangorestframework==3.12.4
whitenoise==5.3.0
psycopg2-binary
RequestsClient
pytz
django-admin-interface
